fix: PPO rollout stores z-score normalized logits — distribution consistency
Root cause: decoder sampled from float32 raw logits (z-score + temp), but
training recomputed log-probs from float16-quantized ±960 logits, producing
a different distribution → old_logp mismatch → ratio/KL explosion.

- rollout: store per-head z-score normalized logits (float16-safe, ~[-3,3])
- ppo_train: old logπ via normalized_log_probs (no re-normalization);
  new logπ via compute_action_log_probs (z-score on raw model logits)
- also: lower default lr 3e-4→1e-4, ppo_epochs 4→2

+++ code/my_ai/ga_ss_boilerplate.py
@@ -108,7 +108,7 @@ def mutate_class_labels(
             logits_i = logits_i + torch.log(w) * temperature  # category rebalancing
             probs = F.softmax(logits_i / temperature, dim=-1)
             sampled = torch.multinomial(probs, 1, generator=rng).squeeze(-1)
-        mutated[mask, hi] = sampled
+        mutated[mask, hi] = sampled.to(mutated.dtype)
 
     return mutated
 
@@ -173,7 +173,7 @@ def mutate_class_labels_soft(
                 mutated_logits[mask, hi, :][row, old_c] = mutated_logits[mask, hi, :][row, new_c]
                 mutated_logits[mask, hi, :][row, new_c] = tmp
 
-        mutated_labels[mask, hi] = sampled
+        mutated_labels[mask, hi] = sampled.to(mutated_labels.dtype)
 
     return mutated_labels, mutated_logits
